+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Populate jobId field on job logs, update `stripes-smart-components` to `v3.0.0` to avoid errors. UIDEXP-35.
 * Populate runBy field on jobs logs and running jobs. UIDEXP-37.
 * Init second pane navigation in settings and link to field mapping profiles section. UIDEXP-39.
+* Add static search form on field mapping profiles settings page. UIDEXP-40.
 
 ## [1.0.0](https://github.com/folio-org/ui-data-export/tree/v1.0.0) (2020-03-13)
 * UI app created. Refs UIDEXP-8.